fix: needs runtime `node-fetch` import
`import ... from "node-fetch"` will throw in Edge runtime and browser
because the library attempts to make `node:...` package imports.

The only way to consume this package is to load it when it's needed with
`await import(...)`, and not as an import statement at the top of the
context.

It is the difference between always loading it and only using it when
needed, and only loading it and using it when needed.

closes https://github.com/SpellcraftAI/openai-streams/issues/50
cc @transmissions11